PythonProject1/requirements.txt
# Selenium WebDriver
selenium==4.15.2

# Data processing and Excel handling
pandas==2.1.3
openpyxl==3.1.2

# JSON handling (built-in with Python, but ensuring compatibility)
# json - built-in module

# Additional utilities
webdriver-manager==4.0.1

# Robot Framework (if not already installed)
robotframework==6.1.1
robotframework-seleniumlibrary==6.2.0

# For better error handling and logging
robotframework-requests==0.9.4
